Remove `Network` trait and use cliquenet directly.
